AMB-1741 move imms api in one place
