--- tests/immunisation_api.py
@@ -1,9 +1,0 @@
-class ImmunisationApi:
-
-    def __init__(self, url, token):
-        self.url = url
-        self.token = token
-
-    def get_event_by_id(self, event_id):
-        # Make your request to our api here and return the response
-        return "response"
