+++ tests/test_immunisation_api.py
@@ -1,6 +1,15 @@
 import pytest
 
-from immunisation_api import ImmunisationApi
+
+class ImmunisationApi:
+
+    def __init__(self, url, token):
+        self.url = url
+        self.token = token
+
+    def get_event_by_id(self, event_id):
+        # Make your request to our api here and return the response
+        return "response"
 
 
 # TODO: send a GET /event/{id} request
